Move test preprocessor to another class so don't need to import matlibplot or other crap
